refactor: removed .env.local, IMap and HealthCheckModel
