--- .env
@@ -1,2 +1,0 @@
-REDIS_HOST=localhost
-REDIS_PORT=6379
