+++ .gitignore
@@ -5,4 +5,4 @@ node_modules/
 .DS_Store
 .vscode/
 .env.*
-!.env
+!.env.local
