--- src/models/HealthCheckModel.ts
@@ -1,4 +1,0 @@
-export interface HealthCheckModel {
-    timesVisited: number;
-    lastTimeVisited: string;
-}
