--- src/utils/Types.ts
@@ -1,1 +1,0 @@
-export type IMap<T> = { [key: string]: T }
